Added GallioTestReportHeader.png to be used in generated test reports.
Changed Gallio-Report.css to use the new image in the header.

Removed the text "Gallio Test Report" from the <h1> tag in Gallio-Report.html+xhtml.xsl (since that same text is contained in the new graphic) but left the empty <h1> tag for now since the css for <h1> has the header image styling information.  (I'll be working on a better way to do the styling, but I just wanted to be able to start using the new graphic without much experimentation.)