Initial checkin of the full integration of the Gallio data driven testing features. We are now fully self-hosting on MbUnit v3 for all tests, including row tests.
Rewrote the PatternTestController.
Added the concept of an IPatternTestHandler and PatternTestActions to support incremental contribution of new test semantics.  The idea is that the PatternAttributes just register actions for the things they want the test to perform.
Added support for static test fixtures.
Added [AssemblyFixture] as a replacement for the MbUnit v2 [AssemblyCleanUp] attribute.
Added support for disposable fixtures.
Added some new conversions.
Added an initial draft of IFormatter, a formatting abstraction.  Will be rewritten shortly to accept pluggable formatting rules.
Made the ExpectedException feature primitive and metadata-based so that it will work magically with any data source that emits the right metadata.
Changed how test step full names are formatted and presented to the user in TDNet and reports.
Added warnings as errors option to MbUnit assembly.  Using #pragmas to disable warnings on old MbUnit v2 code that has yet to be replaced.
Lots of refactoring all over the place.

Note: There is still a bit of work remaining after this checkin, particularly in the way of new tests to be written.