In the process of fixing some bugs in the data binding code, I realized that the Pattern Test Framework had a few deficiencies related to how data sources are scoped. This little excursion led to the following improvements:
Added PatternTestDataContext to capture scoping for data.
Improved the implicit data binding support.  The binding index that was used was wrong in many cases.
Simplifies the evaluation of patterns by refactoring the whole IPatternTestBuilder mess into PatternEvaluator and PatternEvaluationScope.  These are more powerful and easier to use.
Added the concept of an Annotation to associate error, warning and informational messages with code elements.  You can think of these a bit like compiler errors.  They are produced during test enumeration and are persisted in the test model.
Modified the test runners to display annotation messages.
Added a ReSharper "Daemon Stage" to present annotations as error messages in the editor.
Improved how errors in patterns are reported.  They are now handled very uniformly by the PatternEvaluator which will publish a suitable Annotation.
Added a [Name] attribute.
Commented out the [Mixin] attribute since it is unimplemented.
Defined standard attribute targets for different patterns.  This is a more robust design than the previously scattered and inconsistent attribute targets used earlier.
Fixed a bug with missing data conversions for field and property slots.
Added some fairly rigorous tests for different data binding scenarios.

This checkin is a work in progress.  Some tests pass but there's probably a bunch of broken stuff remaining.