Added support for ReSharper 4.0 using mirror projects (because the referenced assemblies are different) and conditional compilation to patch things up. One notable improvement here is the new AttributeValue class in R# 4.0.  This allows us to discard a bunch of hacks we had to use with R# 3.1.
R# build 783: Seems to work fine.
R# build 803: Does not seem to run the tests in the remote task runner process.

Improved the ReSharper handling of tests with multiple results such as data-driven tests.

The build now depends on MSBuild v3.5 due to a requirement to read the registry to find dependencies such as R#, MSTest and Pex.