Hmmph - had problem with connection for some reason. Trying again
First crack at docs for the rest of the Assert, Decorator and Fixture classes in MbUnit.Framework. Have used the opportunity to fold in some of the wiki docs where appropriate and also used some of the nice descriptions Jeff has given to those classes which hav sirvived into v3. Marked a few classes Obsolete as well.

Still a bit uneven, but if someone could rebuild the docs on docs.mbunit.com with this set of additions, many people would be a lto happier I think.