Moved the Gallio.Copy project to the Extensions folder because it's more of an add-on tool than an intrinsic part of the system unlike the Host, Utility and ControlPanel apps in the Gallio folder which provide essential services or directly expose services wholly defined in the core libraries.  (Hope that makes sense, I had to think about it for a bit...)
Modified the Sync VS2010 Projects script to only issue a warning when a VS 2010 project appears to be missing.  This is useful for prototyping spikes for new components that have not yet been integrated into the build.

Added VS 2010 projects for the Gallio.Copy app, changed the output path and enabled warnings as errors per our project convention.