Major refactoring / redesign of the Gallio test object model and test lifecycle.
Added full support for file-based test frameworks in the core.

Test frameworks are now in charge of setting up their own isolation context if desired instead of receiving a ready-made host.  An ITestDriver now refers to an object created by an ITestFramework to drive the process of test exploration and execution.  Basically this turns the whole lifecycle inside out.

Reduced the surface area of the API needed to implement a test framework adapter.  TestModels, TestControllers, TestCommands and TestExplorers are now optional components that a framework's TestDriver may use as part of its implementation.  Previously they were essential parts of the API.

Introduced a Describe method alongside Explore to improve the separation of concerns between reflection-only description and heavyweight exploration in a test isolation context.

Redesigned TestPackage and TestProject.  Among other things they now support an overlay scheme where settings can be partially overridden by another instance.

Moved support for running test projects and wildcard processing into TestLauncher.

Added the concept of a test isolation provider which is responsible for providing an adequate hosting environment just for tests.

Replaced synchronous listeners with asynchronous messaging.  This is still a work in progress.  Later on I plan to add a peer-to-peer message bus and further restructure the communication patterns among components.

Reclassified the Schema types so they appear in namespaces that are subordinate to the modules they support.

Note: This check-in includes many breaking changes.  Notably, test projects and test reports created prior to this check-in will no longer be compatible.