Removed the test framework node from the test tree.
Modified Icarus to fully support test files (instead of just assemblies).  Removed all logic in Icarus that was predicated on the test kind; we should be relying on structural invariants instead.

Added a test for NUnit SetUpFixture and fixed a few bugs related to that and having "AutoNamespaces" enabled.

Added status reporting information to the ProgressMonitor dialog.

Fixed several regressions.