Fixed several issues related to MSTest deployment items and data-driven tests.
Extended AssemblyMetadata to support detecting the version of the .Net runtime that an assembly was compiled for.

Modified the DotNetTestDriver to use the original runtime version of an assembly unless otherwise requested to use something else.  This means .Net 2.0 test assemblies will run with 2.0 by default and .Net 4.0 will run with 4.0 likewise.  Among other things, this helps to resolve some debugging problems because we were promoting test assemblies to the newer runtime but sometimes trying to run them with an older debugger.

Added support for .Net 4.0 unresolved code elements.  We have to use two different copies of these classes because some new API has been introduced.  For example, now we can directly subclass Assembly instead of having to use a shim.

Modified the sample runner to embed XML into the test result to help with diagnostics.