Fixed an issue where the MSTest runner would emit duplicate test finished messages.
Fixed an issue with ObservableTestContext where it was not setting IsFinished to true as it should have.  That would have made the first bug more obvious since it would have caused an exception to be thrown elsewhere.  *sigh*

Added support for displaying "RunInfo" error messages from an MSTest test run.  This is where assembly and class errors get reported.

Fixed some docs typos.