Fixed some bugs in the VS 2010 install that happened because the wrong version of certain services were being registered.
Added a new conditional enablement feature for plugins so that we can more precisely select features that ought to be enabled based on the current context.  For example, when running in VS 2008, one version of the DefaultShell component is registered, in VS 2010 a different version is registered.

Pulled common non VS-specific parts of the Shell interfaces out to a separate project.

Added some code to detect when csUnit is not going to be able to run tests because some of its components are not available in the GAC or in the test assembly's directory.  This is due to a bug in how CSUnit configures its AppDomain and loads the RemoteLoader.  It is using the AppDomainSetup.PrivateBinPath property incorrectly.

Refactored much of the Shell to take advantage of the Gallio v3.1 plugin loading process such as using declarative traits and component handles for lazy loading.

Added some missing security attributes to derived reflection types like UnresolvedAssembly.