Backported rev 2194.
Rearranged the samples projects.
Added new samples for WatiN and White.
Added a Sources.txt file to document the SVN repository Url and revision used for a build.
Upgraded xUnit to 1.5 RTM.
Reenabled the context menu for the test step run viewer.  Should make it easier for users to navigate if necessary.
Performed a couple of minor tweaks to the pattern test framework to facilitate the creation of new kinds of data-driven tests such as those created by the Browser attribute in the WatiN sample.