Fix for issue 518: "AutoCAD integration fails on Windows Vista".
Changed from SendMessage(WM_COPYDATA) to a OLE Automation approach using AutoCAD's "AutoCAD.Application" COM object from the running object table.

There are a couple trade-offs with this fix:
1. If Gallio is creating a new AutoCAD process there can be no existing ones running.
2. If Gallio is attaching to an existing AutoCAD instance there can't be more than one running.

The reason for this is that when AutoCAD registers its "AutoCAD.Application" object with the ROT it doesn't use a moniker unique to the process. If multiple AutoCAD processes are running then only the one that registered itself first can be retrieved using the GetObject API. If AutoCAD included some bit of unique data in the moniker, such as the process ID, then we could uniquely identify the object.