Refactored the build system.
Dropping MetaBuild in favor of a build system that will be somewhat less monolithic.  Modules can be built independently and then composed as required.  The objective is to allow us to partition builds across multiple VMs with different system configurations and to enable the creation of new distributable formats (archives and installers) that include a subset of the modules.

As part of this effort, moved the ControlPanel and Utility projects to Extensions.  Also extracted MbUnit.Compatibility.Tests from MbUnit.Tests.

Generated brand new solution files for all modules.  Each solution contains a minimum number of projects to ease development of individual modules.

This check-in is a work in progress.  It will break the build until the refactoring is complete.